Integrate new abstraction in RegisterActivity & tests
