+++ app/src/main/java/ch/epfl/sdp/kandle/User.java
@@ -10,8 +10,8 @@ public class User {
 
     public User(String id, String username, String email) {
         this.id = id;
-        this.username = username;
-        this.email = email;
+        setUsername(username);
+        setEmail(email);
     }
 
     public String getId() {
